net: ipa: Add support for IPA v2.x in the driver's QMI interface
On IPA v2.x, the modem doesn't send a DRIVER_INIT_COMPLETED, so we have
to rely on the uc's IPA_UC_RESPONSE_INIT_COMPLETED to know when its
ready. We add a function here that marks uc_ready = true. This function
is called by ipa_uc.c when IPA_UC_RESPONSE_INIT_COMPLETED is handled.

Signed-off-by: Sireesh Kodali <sireeshkodali1@gmail.com>
Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>